Create new GPIO header.
Has 14 GPIOs with reasonable availability of peripherals (timers, usarts, ...).

The location is in upper left; USB is moved to the middle left.

(Neither USB nor generic GPIO are expected to be populated on the
actual LED-torus, but could turn out useful for other purposes).

Routing for this is preliminary, needs cleanup.